pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "morpheus"
version = "0.1.0"
description = "Event-driven trading system with meta-labeling and risk overlay"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Morpheus Team" }
]

dependencies = [
    "pydantic>=2.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["morpheus"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py311"
